sys-firmware/intel-microcode: avoid mount-boot if ROOT is set
and:
- rdep on iucode_tool for .install hook
- USE=hostonly: iucode_tool is executed in pkg_preinst and should therefore be
an idep instead of a rdep (not that it really makes a functional difference
since it is bdep anyway)
- rework REQUIRED_USE: the *.install hooks require the split-ucode to be
present. Since with USE=dist-kernel, USE=initramfs does not install intel-uc.img
(instead it is delegared to installkernel via the *.install hooks)  we need
always split-ucode with dist-kernel.
- *.install: exit gracefully if no ucode installed

this mirrors recent changes in sys-kernel/linux-firmware

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>