profiles: update egl-wayland-1.1.14 mask description and widen
Issues seem to be piling up, and while firefox upstream fixed one
issue on their end in nightly, there may be more to it that needs
fixing on egl-wayland's side and so this mask may be there for a while.

Including revbumps so it's not accidentally unmasked in the interim.
Will likely wait for 1.1.15 to consider unmasking even if opt to
backport some fixes in revbumps, unless it takes a while and people
want this earlier.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>