sys-apps/openrc-navi: new package, add 0.60_beta, 9999
this package is a fork of openrc with user services support implemented,
a few current bugs addressed, and other features and bugfixes planned.

after a year of on-off development of the user services feature, little
was heard from openrc's maintainers, since upstream seems rather
stagnated, this fork was made to push for for improvements for the
project and ecosystem as a whole.

Signed-off-by: Anna (navi) Figueiredo Gomes <navi@vlhl.dev>