dev-util/perf: Use cross CLANG for compiler feature verification
The build system of perf checks the clang version in order to set
correct compilation flags. However, it defaults to the non-prefixed
clang invocation, which fails during cross-compilation.

Set CLANG to point to the fully qualified host clang.
The build system will only use it if it’s actually available.

Closes: https://bugs.gentoo.org/936558
Closes: https://github.com/gentoo/gentoo/pull/37688
Signed-off-by: Andrej Shadura <andrew.shadura@collabora.co.uk>
Signed-off-by: Guilherme Amadio <amadio@gentoo.org>