dev-qt/qtmultimedia: update >=6.8 live
Originally thought it added native pipewire audio support, but this
is only for screencast and so not doing USE=pipewire.

Split off [X=] hack and add a redundant [X?] in case we are ever
able to remove it (aka gst not broken without xorg-proto, xorg-proto
in RDEPEND, or if a new EAPI adds a way declare DEPEND-only deps for
reverse deps).

gst[egl] could technically be optional, but feel it's not worth
introducing a USE. Arguably feel it should be unconditionally
enabled on gst if USE=opengl.

Updating the ebuild also exposed a bug in cmake's checks with gl_x11
and gl_wayland, been reported and doing a temporary workaround for now.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>