dev-python/PyQt6: restrict to <qtbase-6.8
Attempted to patch it like usually do, but this time around some
methods were replaced by alternatives that I'm not sure sip even
supports and have no idea how to patch it properly (esp. since
hardly know sip and C++ as-is). Issues are in coreand USE=testlib,
the rest appears fine.

Not much can do other than wait for PyQt6-6.8.0 to release
(similarly to what pyside6 does) unless someone else can provide
a working patch. If there is a PyQt6 pre-release snapshot will
likely use it once available, but these haven't been happening
*before* Qt releases lately.

Technically no need for a revbump thanks to the binding operator,
but, if we do not do it, portage backtracks a lot more and gives
confusing messages when the whole Qt stack is involved. So want
to propagate the upper bound on user's systems now *long* before
the bump (may be wasted if we get a snapshot in time, but that
seems unlikely). Will spare stable users given we can just stabilize
6.7.1-r1 to cause a rebuild and odds are we will not stable 6.8.0
either way (may wait for 6.8.1 + PyQt6-6.8.x).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>