x11-drivers/nvidia-drivers: note reminder for next release
Rather not force the dependency yet to allow easy downgrading,
hopefully getting these out together will work out decently.

Those opting in this beta may want to unmask and try it though.

Supposedly the bundled prebuilt egl-wayland is even worse given
it includes a regression that is not present in source-built
1.1.13 and lacks the fix from 1.1.14.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>