www-client/chromium: add 127.0.6533.72
Drop llvm-r1 eclass for llvm-utils and our "llvm-r1_pkg_config at home".

Chromium is one of the few builds that won't benefit from the default
eclass behaviour and we don't currently need to worry about linking
with llvm libc++.

We'll leave much of the machinery in place an instead take advantage
of llvm-utils to do the checks ourselves and try to enforce our own
dependencies / sanity checks on rust too.

Bug: https://bugs.gentoo.org/936611
Signed-off-by: Matt Jolly <kangie@gentoo.org>