dev-util/bindgen: remove subslot binding to clang runtime dep
(For clarity, force it to bind to :* so no one tries to change this in
the future.)

It is NOT a build dependency. This is an indicator for the fact that it
is not needed to build, because bindgen simply compiles rust code that
knows how to try dlopening various names for libclang.

Consequently, it also doesn't matter which version of clang you have
installed -- any will do (well, technically, any version greater than
clang 6 will do), and in fact it is fully runtime switchable.

There's no reason to bind to whichever version may happen to have been
installed at the time of building. It is definitely not useful to
trigger a subslot rebuild every time clang updates.

Moreover, it would not bind to any subslot at all if you used
--buildpkgonly without clang installed at all. (In fact, this RDEPEND is
a direct violation of PMS, so the fact that it is mishandled isn't
terribly surprising.)

No revbump. The point of this change is so that people don't have to
keep rebuilding pointlessly, so the fact that people who don't use
--changed-deps won't see this change is a desirable feature.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>