media-libs/noise-suppression-for-voice: add 1.10-r1, update live
The new revision installs a sample PipeWire configuration file.
Additionally, we're getting rid of the "ladspa" USE flag, to make sure
the configuration always works. This shouldn't be a problem as the
ladspa plugin is not that heavy compared to all the other ones. This
also lets us clean up REQUIRED_USE.

Closes: https://bugs.gentoo.org/936789
Signed-off-by: Kacper Słomiński <kacper.slominski72@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>