sys-process/glances: add 4.1.2.1
Upstream has migrated from ujson to orjson. orjson in Gentoo doesn't
have pypy3 compat. Also mask >=4.1.2.1 in the wd40 profile as a new
reverse dependency of Rust-requiring orjson.

Also note that here introduces a hack to handle a mismatched release
version with the version in code, and this is hacked around through
src_prepare.

Signed-off-by: John Helmert III <ajak@gentoo.org>