dev-build/meson: disable CUDA tests
If CUDA is installed, we try to test it and everything fails. The issue
seems to be that CUDA simply cannot handle the use of LDFLAGS. It may be
possible for meson to mangle these for you, but it currently does not.
There's a tracking ticket for it.

It's preferable to test common functionality rather than work around
bugs. Just pretend CUDA doesn't exist as far as src_test is concerned.

Closes: https://bugs.gentoo.org/936757
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>