dev-util/ruff: add 0.5.5
I have recently noticed that running cargo test causes the resulting
tested binary to be configured without color support. In general this is
a huge flaw of the cargo "build" system -- it cannot handle building
something, then testing it, so instead it rebuilds the original thing.
It makes me incredibly nervous and clearly it is going wrong somehow --
and this is "expected functionality" in rust land, so plausibly the
canonical solution is to simply run tests as part of a completely
separate build.

It is not like reusing the same build directory will make anything go
faster. cargo's caching is very primitive here and basically all crates
will get recompiled anyway, for one reason or another.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>