dev-python/setuptools: add 72.1.0, drop 72.0.0
In 72.0.0 the test command was finally removed as has been threatened
for a while. It turns out that many people still had referenced it in
their setup.py even if they didn't actually use it, so simply building a
wheel then failed -- the only deprecation warning had been if you
actually attempted to use `setup.py test`.

Setuptools 72.1.0 adds a deprecation shim to allow those packages to
still build wheels successfully, to give projects time and warning that
they need to remove this last remaining trace.

We drop the old version too, as we don't want users to hit the version
that breaks wheel building.

Bug: https://github.com/pypa/setuptools/issues/4520
Bug: https://github.com/pypa/setuptools/pull/4522
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>