dev-python/scipy: remove outdated workaround for pythran
The upstream issue was fixed and released in 0.16, which is also the
only version we package anymore. Bump the dependency to make sure people
actually updated this bdep, and drop our hack.

This also fixes the problem where people misdiagnose a faulty personal
environment as a portage bug, due to setting PYTHONPATH="". Well...
actually it may hide the issue entirely, if scipy.signal as the `signal`
module doesn't actually get used. :) But that is beside the point.

Bug: https://bugs.gentoo.org/936948
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>