x11-drivers/nvidia-drivers: try using source-built egl-gbm again
It *looked* like egl-gbm repo was missing changes, but the single
1 line commit from 3 months ago turned out to be relevant for 560
and egl-gbm hadn't got a release in the interim.

Source egl-gbm is now patched, so let's go back to it before we
keyword a 560 release.

Also cleanup outdated egl-wayland comment, that's resolved thanks
to 1.1.13.1 receiving the same change as egl-gbm.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>