media-sound/lmms: add missing testsuite dependency
It needs dev-qt/qttest in order to link tests successfully. And there is
also a cmake warning that Qt5::Test was linked to, but not found, and
that's a bug -- due to cmake's excellent, world-class type safety, this
is NOT obviously a detectable error when you can simply assume maybe it
means `-lQt5::Test`, but in the minimum version that the live ebuild
uses, the fact that the string name includes "::" means cmake can and in
fact does error out since "::" is reserved for imported libraries.

So the live ebuild actually fails to configure without qttest, but even
the versioned release would error out when running the testsuite and
attempting to compile the test program. Of course, the versioned release
also doesn't *connect* the testsuite, and one of the test cases appears
to fail anyway even if you manually cd and run it by hand.

Naturally, there is no buildsystem option to disable this. Instead, we
use a filthy sed to the test directory when not building tests.

Closes: https://bugs.gentoo.org/879567
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>