net-misc/curl: add 8.9.1; enable HTTP/3 (QUIC) by default
The support for HTTP/3 is stable enough and we haven't seen any
reported issues since unmasking the USE where possible.

HTTP/3 is default-enabled using dev-libs/openssl[quic],
with the USE_EXPAND and REQUIRED_USE enforcing required deps
and enabling users to select an alternative implementation.

The only currently-supported option for this is net-libs/ngtcp2[gnutls],
though it is expected that as other TLS backends implement QUIC
(or are supported by other backends) that this will increase.

Of note is that HTTP/3 support and MultiSSL are mutually exclusive at
this point in time. There does not appear to be a significant benefit
to MutliSSL in 2024:

> An original driving factor for this feature was to allow Schannel with
> other TLS backends on Windows so that users could access "native" CA certs
> in the Windows CA store. Subsequently, curl has introduced support in multiple
> TLS backends to use native CAs.
> This reduces the need for many people to use and switch TLS backends.
> There are probably also other use cases, since the TLS backends are not all alike.

We're not dropping MutliSSL support, however HTTP/3 as the default seems
like the option that provides the greater utility as a default
configuration.

This is not to say that MultiSSL support is inherently incompatible with
HTTP/3, it will just require a sufficiently motivated developer
to implement the code.

Closes: https://bugs.gentoo.org/936627
Signed-off-by: Matt Jolly <kangie@gentoo.org>