dev-libs/openssl: enable USE=quic by default.
net-misc/curl enabled HTTP/3 (QUIC) by default in
00f5031e36ffde7784d10ded6f2c753c3a5513a5.

As OpenSSL is the default TLS backend for cURL in
Gentoo, is a well-maintained and tested package,
and cURL's multiple-tls backend support (MultiSSL)
is not currently supported with HTTP/3, it has been
selected as the default QUIC backend for cURL in Gentoo.

Enable USE="+quic" to simplify dependency resolution
for the majority of consumers.

Signed-off-by: Matt Jolly <kangie@gentoo.org>