toolchain-funcs.eclass: Fix tc-is-lto not to leave stray files in T
Fix tc-is-lto function to remove the temporary file after testing.
Besides being cleaner, this fixes a permission problem when using
Paludis and tc-is-lto is used both in pkg_setup() (which creates
the temporary file owned by root) and src_*() phase (which attempts
to rewrite it as a regular user).

Thanks to negril for the report!

Signed-off-by: Michał Górny <mgorny@gentoo.org>