dev-vcs/git: Fix cross-compiling when build host lacks curl
The Makefile tries to run curl-config, which is practically always present on
the build host, but we shouldn't assume that, and it could return the wrong
values when cross-compiling.

It's just a shell script, so we can safely call the copy under ESYSROOT for the
version check. For the CFLAGS and LDFLAGS, leverage pkg-config instead, as it
will correctly adjust the paths.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>