dev-db/percona-xtrabackup: fix automagic dependencies
There were some dependencies which we tried to link to, which are
usually bundled but we ask for system copies. However, we didn't
properly depend on it -- and in the case of procps, even though it is in
@system we do want to enforce a subslot dependency on it, so it really
shouldn't be left out.

rapidjson doesn't actually work. You can try to unbundle it but it will
simply error out because rapidjson hasn't been released since 2016
(despite active development, they simply ghost anyone requesting a
release) and percona states they are using newer features... there is an
error check if you try to force system rapidjson and whatever they need
isn't present.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>