dev-db/percona-xtrabackup: add 8.3.0.1
Bump to latest upstream release. This is years out of date,
maintainer-needed, and patched for build errors that are fixed upstream.

I'm relying on the fact that other distros package this update without
issue. On which note: other distros also package some files we don't! So
go ahead and package those.

The root cause is probably this rather hackish src_install. The upstream
install routine is quite bad as it installs loads of stuff nobody wants
-- some distros just run cmake install and then delete the lot instead.
The software comes with especially interesting gems such as attempting
to install the manpages we do include, to ${ED}/var/tmp/portage/.....

I figure that globbing the plugins for manual installation is probably
the safest move, all things considered.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>