www-client/firefox: add support for ppc64le on musl
It still does not add support for ppc64 (big-endian) because the USE
flag "ppc64" is shared between ppc64 and ppc64le.

Fixing it on ppc64 would probably require adding a big-endian USE and
there are some big-endian specific patches necessary. Since I don't have
yet a big-endian Gentoo, this patch does not fix ppc64 on musl, but it
currently does not work anyway.

Signed-off-by: Piotr Kubaj <pkubaj@anongoth.pl>
Closes: https://github.com/gentoo/gentoo/pull/38018
Signed-off-by: Joonas Niilola <juippis@gentoo.org>