gui-libs/egl-wayland: add 1.1.15
There may still be issues but situation should be a lot better than
it was with 1.1.14, so trying to add this unmasked.

If needed, one can (newly) set __NV_DISABLE_EXPLICIT_SYNC=1 to disable
that rather than downgrade to 1.1.13.1.

Also note that explicit sync is not enabled unless use the (currenty)
unkeyworded 560 drivers. As far as I can tell 1.1.15 does nothing
special if used with older drivers and works normally.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>