app-misc/tmuxp: fix broken test dependency on typing-extensions
It is needed for all versions of python, not simply arbitrarily 3.10.
Although it actually should not be needed -- the issue is that pytest
automatically tries to import all files in tests/ and scan them for
testcases. Even a file that exists solely to be used by mypy and has no
tests, still gets imported for collection purposes.

Instead, just make the fixtures be ignored by the collector. Now
typing-extensions isn't needed at all.

Closes: https://bugs.gentoo.org/934083
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>