games-simulation/flightgear: use https for git clones in live
Using git:// is a really bad idea and the eclass yells at you for it.
And the second uri does not actually exist (anymore).

Note: the live ebuild still doesn't actually build. The patch is stale
and then there are a bunch of weird "XXX does not name a type" and
"YYY was not declared in this scope" errors that I have no patience to
debug. Perhaps an upstream bug. I was able to determine that existing
LTO errors are still present and then gave up. Baby steps... :D

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>