sys-cluster/torque: mark as LTO-unsafe
The project appears to have gone source-unavailable and all-in on AI.
There are 6.x versions available mysteriously as some git branch with no
distfiles (Fedora packages this) and a 7.x version is announced on their
website but you have to contact sales to find out more. They haven't
communicated on github in years.

Not only is reporting a bug an exercise in futility, I'm not super
interested in trying to begin with.

Closes: https://bugs.gentoo.org/863737
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>