sci-libs/opencascade: fix missing libGL
This is a due-diligence update of the dependency on media-libs/libglvnd.
There are no other known reports of this issue outside of automated testing.

Given that this ebuild already fails on systems where media-libs/libglvnd[X]
isn't already pulled by other packages it seems to be the better choice to avoid
the huge rebuild and refrain from rev-bumping.

Closes: https://bugs.gentoo.org/936901
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/38121
Signed-off-by: Sam James <sam@gentoo.org>