profiles/arch: update QUIC/HTTP/3 mask for curl
As most arches now meet the requirements to enable
HTTP/3 via one or more of the available QUIC backends,
update so that we're masking on arches without support
rather than unmasking where there is.

This means that we're masked on: alpha, ia64, ppc, ppc64.

I'll fire off appropriate keyword requests to enable this
support where it makes sense.

Signed-off-by: Matt Jolly <kangie@gentoo.org>