dev-lang/yap: do not require python_setup to run if not USE=python
We don't need to check for a python impl, since python itself is
optional. This is especially relevant given that it hasn't been migrated
to python 3.12, so by default there is no available implementation
selected, and python support is also NOT built by default, so this
change makes it build by default.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>