sys-devel/bpf-toolchain: add USE=bin-symlinks
Note that unlike dev-util/mingw64-toolchain, we're enabling it by default
because it'll be pretty uncommon for anyone to have a crossdev-built
toolchain for it. This is needed to avoid conflicts if someone wants
to build a bpf-unknown-none toolchain via crossdev where an ebuild
depends on bpf-toolchain (wouldn't be possible before because of file
conflicts).

Also, with it off by default, we would have to add some hacks in e.g.
the systemd ebuild to make it work by mangling PATH and things get
tricky there especially given only a small part of the build would even use it.

May change mind on this later if needed though.

Signed-off-by: Sam James <sam@gentoo.org>