net-misc/trurl: add latest 0.14, fix 9999 build
The old Makefile patch can no longer be applied; so I wrote a new patch
that applies to 0.14 that is also applied to the 9999 version for now,
and sent it upstream.

Additionally, I added an  export NDEBUG=1  command in src_compile:
without this, -g and -Werror are added after CFLAGS by the Makefile
possibly shadowing options like -ggdb3 that user might have set, or
causing a build to fail just because a warning was reported.

Bump copyright of touched ebuild files.

Ref: https://github.com/curl/trurl/pull/322
Signed-off-by: Emanuele Torre <torreemanuele6@gmail.com>