dev-python/pyocr: remove support for USE=cuneiform
It defaults to tesseract anyway, and cuneiform hasn't been maintained
since 2011 and generally doesn't build. Unlikely anyone has any desire
to use it. Upstream only continues to support cuneiform at all because
they want to "exercise the code" with regard to its support for multiple
runtime-switchable backends.

cuneiform is being last-rited. In the unlikely event someone was using
it, they will be warned to uninstall it and install tesseract instead by
the mask. If they really don't want to, they can ignore the mask *and*
the pyocr rebuild. In the event that people were using tesseract, they are
perfectly fine anyway and don't need to rebuild (but changed-deps
and changed-use will both try to rebuild). No revbump.

Bug: https://bugs.gentoo.org/937994
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>