dev-python/sip: update licenses
Like several other packages from the same upstream, this
changed to BSD-2 after its introduction on github but I
seem to have overlooked updating it for this one.

Considering how development been semi-closed prior to github
and how a lot of the sip code been rewritten recently, there
is "probably" no issue with this change.

Closes: https://bugs.gentoo.org/938123
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>