dev-python/pythran: mark as incompatible with setuptools 73
setuptools has a high rate of churn and now tends to frequently break
consumers that subclass or otherwise use the distutils API in complex
ways. Today's incompatibility, in particular, happens due to metaclass
incompatibilities introduced in setuptools.

Pythran fails hard with this (unstable) version of setuptools. Prevent
the two from being installed together while we wait for a fix to be
determined.

Bug: https://github.com/pypa/setuptools/pull/4503
Bug: https://github.com/serge-sans-paille/pythran/issues/2228
Bug: https://bugs.gentoo.org/938246
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>