gnome-extra/cinnamon: correct USE dependency on gtk
Actually, it doesn't use gtk's gdk-wayland backend at all, not even
optionally, so the dependencies are effectively runtime-only. On the
other hand, it does use the X11 one! Make sure that it correctly
requires gtk+[X] when built for X11 -- which is always, as it currently
doesn't support the wayland backend in that bit of code at all. Hence
wayland support being masked at the moment. So we simply don't worry
about that at all for now. :)

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>