media-video/obs-studio: add 30.2.3
This restores "legacy ffmpeg" being used via USE="nvenc", which
seems to be more in line with the upstream default for the time
being, and there seem to be potential performance problems [1]
with the new, "native nvenc", so stick to "legacy" for now.

The next non-patch release should have CMake options tweaked to
handle these better (will be in 9999 soon), and if there is
want for experimenting more with it already, can always flip
it using 'MYCMAKEARGS="-DENABLE_NATIVE_NVENC=ON"'.

1. https://github.com/obsproject/obs-studio/issues/11009

Closes: https://bugs.gentoo.org/936826
Closes: https://bugs.gentoo.org/938022
Closes: https://github.com/gentoo/gentoo/pull/37755
Signed-off-by: Jimi Huotari <chiitoo@gentoo.org>