app-text/calibre: add 7.17.0
There have been some updates to the binary builds for Qt stuff that
basically amount to building PyQt6 with USE=speech. It's not yet active
in live codepaths, so patching it out of the testsuite for now. Will
keep an eye on it in the future with attention paid to the
practicalities of USE=speech. Recheck in future version bumps...

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>