app-emulation/libvirt: Depend on sys-libs/libnbd too
In its upstream commit of v9.8.0-rc1~27 [1] libvirt introduced an
optional dependency on libnbd. Reflect this in corresponding
ebuilds. Now, pre-10.0.0 ebuilds do not have nbd USE flag. Thus,
it needs to be introduced to record the dependency properly.
NB, initially, the configure option was named 'libnbd' and only
after v9.9.0-rc1~108 [2] it was changed to 'nbdkit'. Hence the
discrepancy in libvirt-9.8.0-r3 and libvirt-9.9.0-r3 ebuilds.

1: https://gitlab.com/libvirt/libvirt/-/commit/40935b395c91a95a71cd66b9e2872346ac3a2145
2: https://gitlab.com/libvirt/libvirt/-/commit/7cbd8c42305735375b60abf8abc47838a7a050d4
Closes: https://bugs.gentoo.org/938346
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/38274
Signed-off-by: Sam James <sam@gentoo.org>