dev-libs/cyrus-sasl: Fix empty sasldb2 generation when cross-compiling
This changes how the generation works as it now needs to be done in src_install
while the native saslpasswd2 is still around, rather than in pkg_postinst. We
cannot check EROOT during src_install, so we generate the empty file
unconditionally. It is easiest to subsequently install this file unconditionally
too but as sasldb2-empty. It is only copied to sasldb2 if that doesn't already
exist. Keeping an empty file around is potentially useful anyway.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>