dev-qt/qtbase: prevent automagically building against gtk[X,wayland]
e.g. the upstream code checks whether GDK_WINDOWING_WAYLAND is defined
by the gtk headers, and if so will compile against the wayland symbols
it provides. This means that qtbase built on a system with
gtk+[-wayland] will be compatible with anything, but when built on a
system with gtk+[wayland], requires that at runtime.

It's only needed, of course, by the qgtk3 platformtheme plugin which not
necessarily everyone builds, wants, or uses at runtime :) but still.

We can solve this with a (new) gentoo-specific macro that hides the
GTK implementations when we don't want it. This is gated on our existing
USE flags for wayland/X support.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>