dev-lang/julia: mark as LTO-unsafe
julia has many problems, at many levels. Among these, it fails to build
with LTO due to various errors somewhat more inventive than the norm
when it comes to LTO issues. Iteration time on building and testing it
is, of course, horrifically bad as it bundles its own private LLVM.

I won't even try to run the testsuite. I tried that years ago on Arch.
"Flaky" doesn't even begin to describe it -- an actual passing test
result sometimes took me 7 builds in a row, and one is never sure why it
actually passed for once. "Fortunately", the ebuild restricts tests
anyway. Needless to say, who knows if it actually works.

It seems to build. Probably. At least better than when -flto is used, so
that's a win. On which note, it is worth pointing out: the ebuild
*tried* to filter lto, but didn't do so early enough. This change simply
lifts the filter early enough that it runs before a sed inserts CFLAGS
into a Makefile, so that the filter actually takes effect.

Bug NOT reported upstream. Julia is a language of all time, and
reporting bugs gets you nowhere except the receiving end of a thorough
gaslighting. I do not volunteer for this. The software is most likely
broken (LTO warnings don't indicate a flawless codebase) but I simply
don't care -- it's not my problem if this programming language is
broken, as long as it isn't an obstacle to other, more useful packages,
being safe to build by default with LTO enabled in make.conf.

Closes: https://bugs.gentoo.org/855602
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>