media-libs/x265: allow building with c++17, small ebuild fixes
The new x265-3.5-r5-cpp-std.patch disabled setting c++98 or c++11 as versions.
This made newer compilers default to c++17 and thus `register` a unused and
reserved keyword.

While on it fix some obvious things that might cause problems:
	one shadowed variable
  use std::abs	in place of abs so we don't accidentally truncate values
	use boolean instead of bitwise comparison when comparing bool

Move ENABLE_CLI=OFF for non native multilib builds to make it clearer.
Don't set ENABLE_TESTS=no on platforms where we set ENABLE_ASSEMBLY=OFF, because
it requires it to be ON and this the option is never created.

Closes: https://bugs.gentoo.org/938515
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>