dev-db/sqlite: Fix build issue in the test suite when USE='-tcl'
Test phase will errors out starting with commit
https://github.com/sqlite/sqlite/commit/21363ac78df6751655c33372a7277512531b9570
as we set HAVE_TCL to 0 in the test phase and with the added check the
build aborts regardless of the fact that 'tclConfig.sh' was found
by the configure script previously. We can avoid setting the variable
manually and the build will suceed with the intended behavior kept.

Signed-off-by: Jakov Smolić <jsmolic@gentoo.org>