net-wireless: Add packages sdrpp, rtl-sdr-blog and its dependent virtual
Add package sdrpp (SDR++).  Add package rtl-sdr-blog, which contains support
for rtl-sdr-blog-V4 hardware.  This package is a fork of rtl-sdr, which
does not have support for this new hardware.  Since rtl-sdr and
rtl-sdr-blog occupy the same slot, new virtual package rtl-sdr-drivers
allows package sdrpp to depend on it to get whichever of the two driver
implementation packages is appropriate.

Signed-off-by: Joe Harvell <landshark@clown-world.org>