dev-java/openjdk: mark as LTO-unsafe
Temporarily disable the USE=lto support for the 21.x ebuild only, as
that's the version it broke in.

OpenJDK cannot get their act together, so their ticketing system is
that you cannot report a bug unless you're a project member, and your
alternative is to fill in a form on the Oracle website and they will
analyze your submission and file a bug on your behalf at some point.
Hopefully.

This is not a ringing endorsement of OpenJDK as an *open* implementation
of Java, but it is what it is... I can't currently provide a public
ticket number but perhaps sometime in the future.

Closes: https://bugs.gentoo.org/916735
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>