dev-libs/gobject-introspection: fix broken update that didn't copy latest
In commit 87757767aae114d32059400fd72d1a59656f8382, the package was
updated with an inaccurate py3.12 bump that wasn't thoroughly tested and
didn't check the changelog. The previous version of the ebuild had
already been updated in the meantime to support py3.12 correctly, via
commit cc2b9a4a8fb4d94acf59dfeb2836cde223415a73.

Copy over the changes from 1.78.1 to 1.80.1.

Fixes: 87757767aae114d32059400fd72d1a59656f8382
Fixes: https://github.com/gentoo/gentoo/pull/36834
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>