sys-auth/oath-toolkit: fix the build on musl
Add an upstream patch to fix the build on musl, and then fix a new
problem introduced by the patch: the build system is extremely sensitive
to mtime changes in the source tree, in that patching triggers a
rebuild of some Makefile sources, which ultimately tries to invoke
automake. We have to fudge the mtime on the patched file to avoid this.

Bug: https://bugs.gentoo.org/936309
Closes: https://bugs.gentoo.org/933086
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>