dev-texlive/texlive-latexextra: move dev-tex/minted to PDEPEND
The RDEPEND in texlive-latexextra on dev-tex/minted was a convenience
dependency because minted used to be bundled.

Unfortunately, we get a cycle:
```
(dev-tex/minted-2.9:0/0::gentoo, ebuild scheduled for merge) depends on
 (dev-texlive/texlive-latexextra-2023_p69752-r4:0/0::gentoo, ebuild scheduled for merge) (runtime)
  (dev-tex/minted-2.9:0/0::gentoo, ebuild scheduled for merge) (runtime)
```

Drop the minted dep to PDEPEND given we don't actually need it
for texlive-lateextra to build, it's just a convenience dep.

Closes: https://bugs.gentoo.org/935405
Signed-off-by: Sam James <sam@gentoo.org>