dev-util/maturin: force shared build & fix tests for musl
While it'd be nice to propagate the shared build, can skip the revbump
(that'd be useless for glibc users given already default) given it
either failed to build, or segfaulted while generating completion
files. If it was successfully installed then it should be usable.

patchelf could be skipped if we disable auditwheel (like distutils-r1
does), but not aware of an easy way to pass this for tests and it
may confuse them either way -- so just add the cheap dependency.

Closes: https://bugs.gentoo.org/938847
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>