dev-libs/gobject-introspection: Require <dev-python/setuptools-74
Force <dev-python/setuptools-74 to fix incompatibility due to removed
deprecated distutils.msvccompiler.  This is a temporary solution
to avoid patching stable versions.  Incidentally, this also extends
the setuptools dependency to older Python versions, but that should
not be a real issue, and we should probably prefer having it for
consistency anyway.

Signed-off-by: Michał Górny <mgorny@gentoo.org>