sci-libs/coinor-dylp: mark as LTO-unsafe
It fails to compile correctly, tests fail, and the resulting library
fails to link when other packages try to use it. The reason is because
it checks for a SunOS symbol using an ancient autotools version that
relies on UB to detect whether it exists, then tries to link to it.

Upgrading autotools is complicated and the software is deprecated and in
maintenance mode, so no telling when it may be released. There are
commits in git master to update it but they are intermingled with
updating to unreleased dependency versions?

Closes: https://bugs.gentoo.org/862696
Closes: https://bugs.gentoo.org/878141
Closes: https://bugs.gentoo.org/878143
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>